services/tasks/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
psycopg2-binary==2.9.1
python-jose==3.3.0
httpx==0.19.0
